tests: move tests to be back under test/ dir to match existing convention
