+++ test/testSerializer.test.ts
@@ -1,5 +1,5 @@
 import pako from "pako";
-import { WARCRecord, WARCParser, WARCSerializer } from "../lib";
+import { WARCRecord, WARCParser, WARCSerializer } from "../src/lib";
 
 const decoder = new TextDecoder("utf-8");
 const encoder = new TextEncoder();
